# api/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "Quiz API",
    "description": "Manage Quizzes"
  },
  "servers": [
    {
      "url": "http://localhost:4022/api"
    }
  ],
...
  "securityDefinitions": {},
  "paths": {
...
    "/quizzes": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Get all quizzes",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "How many elements to skip. Used for paging.",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "How many elements to take. Used for paging.",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort by field",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "in": "query",
            "description": "Search by name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "description": "Search by list of statuses. Example: 1,2,3",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topics",
            "in": "query",
            "description": "Search by list of topics. Example: 1,2,3",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ids",
            "in": "query",
            "description": "Search by list of IDs. Example: 1,2,3",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quizzes fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSearchResult"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSearchResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "read:quizzes"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Quiz"
        ],
        "description": "create a quiz with questions",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Quiz created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "create:quizzes"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddQuiz"
              }
            }
          }
        }
      }
    },
    "/quizzes/{quizId}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Get one quiz including questions",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizExtended"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/QuizExtended"
                }
              }
            }
          },
          "404": {
            "description": "Quiz not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "read:quizzes"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Quiz",
          "Status"
        ],
        "description": "Delete a quiz",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz deleted successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "delete:quizzes"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Quiz"
        ],
        "description": "Update a quiz",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz status updated successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "update:quizzes"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuiz"
              }
            }
          }
        }
      }
    },
    "/quizzes/{quizId}/status": {
      "put": {
        "tags": [
          "Quiz",
          "Status"
        ],
        "description": "Update status of a quiz",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "update:quizzes"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuizStatus"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "auth0": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://dev-ki47hyci.us.auth0.com/authorize",
            "scopes": [
              "read:quizzes",
              "create:quizzes"
            ]
          }
        }
      }
    },
    "schemas": {
...
      "Quiz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Quiz"
          },
          "topicId": {
            "type": "number",
            "example": 1
          },
          "statusId": {
            "type": "number",
            "example": 1
          },
          "createdById": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "example": "2020-03-31T00:00:00.000Z"
          }
        },
        "xml": {
          "name": "Quiz"
        }
      },
...
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "text": {
            "type": "string",
            "example": "what?"
          },
          "choice0": {
            "type": "string",
            "example": "one"
          },
          "choice1": {
            "type": "string",
            "example": "two"
          },
          "choice2": {
            "type": "string",
            "example": "three"
          },
          "choice3": {
            "type": "string",
            "example": "four"
          },
          "correct0": {
            "type": "number",
            "example": 1
          },
          "correct1": {
            "type": "number",
            "example": 0
          },
          "correct2": {
            "type": "number",
            "example": 0
          },
          "correct3": {
            "type": "number",
            "example": 0
          },
          "createdBy": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "example": "2020-03-31T00:00:00.000Z"
          }
        },
        "xml": {
          "name": "Question"
        }
      },
...
      "QuizExtended": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Quiz"
          },
          "topicId": {
            "type": "number",
            "example": 1
          },
          "statusId": {
            "type": "number",
            "example": 1
          },
          "createdById": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "example": "2020-03-31T00:00:00.000Z"
          },
          "questions": {
            "xml": {
              "name": "questionarray"
            },
            "$ref": "#/components/schemas/QuestionArray"
          }
        },
        "xml": {
          "name": "QuizExtended"
        }
      },
      "AddQuiz": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Quiz"
          },
          "topicId": {
            "type": "number",
            "example": 1
          }
        },
        "xml": {
          "name": "AddQuiz"
        }
      },
      "UpdateQuiz": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Quiz"
          },
          "topicId": {
            "type": "number",
            "example": 1
          }
        },
        "xml": {
          "name": "UpdateQuiz"
        }
      },
      "UpdateQuizStatus": {
        "type": "object",
        "properties": {
          "statusId": {
            "type": "number",
            "example": 1
          }
        },
        "xml": {
          "name": "UpdateQuizStatus"
        }
      },
      "QuizArray": {
        "type": "array",
        "items": {
          "xml": {
            "name": "quiz"
          },
          "$ref": "#/components/schemas/Quiz"
        },
        "xml": {
          "name": "QuizArray"
        }
      },
      "QuizSearchResult": {
        "type": "object",
        "properties": {
          "search": {
            "type": "string",
            "example": "query string"
          },
          "topics": {
            "type": "string",
            "example": "1,2,3"
          },
          "statuses": {
            "type": "string",
            "example": "1,2,3"
          },
          "sort": {
            "type": "string",
            "example": "name"
          },
          "order": {
            "type": "string",
            "example": "asc"
          },
          "skip": {
            "type": "number",
            "example": 0
          },
          "take": {
            "type": "number",
            "example": 10
          },
          "total": {
            "type": "number",
            "example": 100
          },
          "items": {
            "type": "array",
            "items": {
              "xml": {
                "name": "quizarray"
              },
              "$ref": "#/components/schemas/QuizArray"
            }
          }
        },
        "xml": {
          "name": "QuizSearchResult"
        }
      },
      "QuestionArray": {
        "type": "array",
        "items": {
          "xml": {
            "name": "question"
          },
          "$ref": "#/components/schemas/Question"
        },
        "xml": {
          "name": "QuestionArray"
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": [],
      "auth0": []
    }
  ]
}